fix: Only create a new space when assigning one in index.tsx. (#13)
Problem: If you load replicache-todo into a tab, let it run, then
delete the backend database, then refresh, the app gets into a
weird state: new mutations are applied optimistically, but aren't
replicated to other clients.

This is caused because pushing to replicache-todo implicitly creates
the "space" row in the database if it doesn't exist. As a result if
an app stays open with a spaceID in the URL while the db is deleted,
that space is recreated with initial values (cookie: 0) when the
server reboots.

From the server's point of view, the client keeps sending mutations
from the future (because the client is at some lmid > 0), so the
mutations are ignored. From the client's pov, the sever just never
accepts any of its mutations, so it keeps trying to resend them.

If the client is refreshed, something even worse happens: a new
clientID is assigned, but the cookie and old client view are kept.
So now the server starts *accepting* mutations on the new space,
but the cookie the client has is from the future from the server's
pov. So every push is accepted, but every pull is a no-op and the
old client view from before the server was rebooted remains.

The core problem here is that Replicache has an invariant
that servers don't go back in time. If a server serves state X with
cookie C, it must never revert to an earlier state with cookie B.
If this happens, existing clients who have cookie C will see
mutations "disappear" because those mutations will advance the
server from B->C, but when those existing clients ask for a diff
from C, they will receive a no-op diff.

Because replicache-todo was implicitly creating spaces that were
not known on push, deleting a space was equivalent to sending it
back in time.

Solution: The solution is to stop implicitly creating spaces on
push. Now they are only created when they are actually allocated
in index.tsx. Push and pull now error if they receive a message
for a space that doesn't exist. As a convenience for developers,
[id].tsx also checks for this situation and redirects back to
index.tsx if it occurs, which causes a new space to be picked.